Pull the pipeline stages out into different compilation units, fix interleaving
